{/src/main/java/com/Loan.java}

+++ src/main/java/com/Loan.java
@@ -16,7 +16,7 @@ public class Loan implements java.io.Serializable {
 
 	private java.lang.Double monthlyRepayment;
 
-	private java.lang.Double interestRate;
+	private Integer interestRate;
 
 	public Loan() {
 	}
@@ -53,17 +53,17 @@ public class Loan implements java.io.Serializable {
 		this.monthlyRepayment = monthlyRepayment;
 	}
 
-	public java.lang.Double getInterestRate() {
+	public java.lang.Integer getInterestRate() {
 		return this.interestRate;
 	}
 
-	public void setInterestRate(java.lang.Double interestRate) {
+	public void setInterestRate(java.lang.Integer interestRate) {
 		this.interestRate = interestRate;
 	}
 
 	public Loan(java.lang.Boolean loanAgeApproval, java.lang.Double amount,
 			java.lang.Integer duration, java.lang.Double monthlyRepayment,
-			java.lang.Double interestRate) {
+			java.lang.Integer interestRate) {
 		this.loanAgeApproval = loanAgeApproval;
 		this.amount = amount;
 		this.duration = duration;
